PR review

+++ ui-tests/cypress/integration/content/assets.spec.js
@@ -4,15 +4,14 @@ describe('Assets', () => {
 
   beforeEach(() => {
     cy.wrap(null).as('assetToBeDeleted');
+
     cy.kcAPILogin();
     cy.kcUILogin('login/admin');
   });
 
   afterEach(() => {
     cy.get('@assetToBeDeleted').then(assetToBeDeleted => {
-      if (assetToBeDeleted !== null) {
-        cy.assetsController().then(controller => controller.deleteAsset(assetToBeDeleted.id));
-      }
+      if (assetToBeDeleted !== null) cy.assetsController().then(controller => controller.deleteAsset(assetToBeDeleted.id));
     });
 
     cy.kcUILogout();
@@ -20,42 +19,28 @@ describe('Assets', () => {
 
   describe([Tag.GTS], 'Add asset', () => {
 
-
     it('Add asset', () => {
-
       openAssetsPage()
           .then(page => page.getContent().openAddAssets())
-          .then(page => {
-            page.getContent().selectFiles('cypress/fixtures/upload/image1.JPG');
-            page.getContent().getGroupSelect().select('administrators');
-            page.getContent().submit();
-          })
-          .then(page => {
-            page.getContent().getAssetsBody().then(rows =>
-                cy.wrap(rows).children(htmlElements.div).should('contain.text', 'image1.JPG')
-            );
-          })
-          .then(() => {
-            cy.assetsController()
-              .then((controller) => {
-                controller.getAssetsList()
-                          .then((response) => {
-                            const {body: {payload}} = response;
-                            cy.wrap(payload[0]).as('assetToBeDeleted');
-                          });
-              });
-          });
+          .then(page => page.getContent().selectFiles('cypress/fixtures/upload/image1.JPG'))
+          .then(page => page.getContent().getGroupSelect().then(select => page.getContent().select(select, 'administrators')))
+          .then(page => page.getContent().submit())
+          .then(page =>
+              page.getContent().getAssetsBody().then(rows =>
+                  cy.wrap(rows).children(htmlElements.div).should('contain.text', 'image1.JPG')));
+      cy.assetsController().then(controller =>
+          controller.getAssetsList().then(response =>
+              cy.wrap(response.body.payload[0]).as('assetToBeDeleted')));
     });
 
   });
 
   describe('Operations on existing asset', () => {
 
-    beforeEach(() => {
-      cy.assetsController()
-        .then(controller => controller.addAsset(testFileInfo, testMetadata))
-        .then(response => cy.wrap(response.payload).as('assetToBeDeleted'));
-    });
+    beforeEach(() =>
+        cy.assetsController()
+          .then(controller => controller.addAsset(testFileInfo, testMetadata))
+          .then(response => cy.wrap(response.payload).as('assetToBeDeleted')));
 
     describe([Tag.GTS], 'Delete asset', () => {
 
@@ -63,21 +48,16 @@ describe('Assets', () => {
         openAssetsPage()
             .then(page => page.getContent().getKebabMenu(testFileInfo.name).openDropdown().clickDelete())
             .then(page => page.getContent().submit())
-            .then(page => {
-              page.getContent().getAssetsBody().then(rows =>
-                  cy.wrap(rows).children(htmlElements.div).should('not.contain.text', 'image1.JPG')
-              );
-            });
+            .then(page => page.getContent().getAssetsBody().then(rows =>
+                cy.wrap(rows).children(htmlElements.div).should('not.contain.text', 'image1.JPG')));
         cy.wrap(null).as('assetToBeDeleted');
-
       });
 
       it('Delete asset referenced by a content - not allowed', () => {
-        let content;
-        let isForbidden = true;
+        const isForbidden = true;
 
         cy.get('@assetToBeDeleted').then(assetToBeDeleted =>
-            content = {
+            ({
               description: 'test',
               mainGroup: 'administrators',
               typeCode: 'BNR',
@@ -99,25 +79,22 @@ describe('Assets', () => {
                   }
                 }
               ]
-            }
-        );
-        cy.contentsController()
-          .then(controller => controller.addContent(content))
-          .then(response => content.id = response.body.payload[0].id);
-        cy.contentsController().then(controller => controller.updateStatus(content.id, 'published'));
+            }))
+          .then(content => {
+            cy.contentsController()
+              .then(controller => controller.addContent(content))
+              .then(response => content.id = response.body.payload[0].id);
+            cy.contentsController().then(controller => controller.updateStatus(content.id, 'published'));
 
-        openAssetsPage()
-            .then(page =>
-                page.getContent().getKebabMenu(testFileInfo.name).openDropdown().clickDelete(isForbidden)
-            )
-            .then(page =>
-                page.getContent().getAlertText().should('have.text', 'Sorry. You cannot perform this action right now, you must first unlink the following cross-references.')
-            );
+            openAssetsPage()
+                .then(page => page.getContent().getKebabMenu(testFileInfo.name).openDropdown().clickDelete(isForbidden))
+                .then(page => page.getContent().getAlertText().should('have.text', 'Sorry. You cannot perform this action right now, you must first unlink the following cross-references.'));
 
-        cy.contentsController().then(controller => {
-          controller.updateStatus(content.id, 'draft');
-          controller.deleteContent(content.id);
-        });
+            cy.contentsController().then(controller => {
+              controller.updateStatus(content.id, 'draft');
+              controller.deleteContent(content.id);
+            });
+          });
       });
 
     });
@@ -127,92 +104,65 @@ describe('Assets', () => {
       it('Edit description', () => {
         openAssetsPage()
             .then(page => page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
-            .then(page => {
-              page.getContent().getDescriptionInput().clear().type('test');
-              page.getContent().submit();
-            })
-            .then(page => {
-              page.getContent().getAssetsBody().then(rows =>
-                  cy.wrap(rows).children(htmlElements.div).should('contain.text', 'test'));
-            });
-
+            .then(page => page.getContent().getDescriptionInput().then(input => {
+              page.getContent().clear(input);
+              page.getContent().type(input, 'test');
+            }))
+            .then(page => page.getContent().submit())
+            .then(page =>
+                page.getContent().getAssetsBody().then(rows =>
+                    cy.wrap(rows).children(htmlElements.div).should('contain.text', 'test')));
       });
 
       it('Crop image', () => {
         openAssetsPage()
-            .then(page => {
-              page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit()
-                  .then(page => {
-                    page.getContent().getImageHeight().invoke('text').then(OriginalImageHeight => {
-                      cy.get('@currentPage')
-                        .then(page => {
-                          page.getContent().openDropDown().openEdit();
-                        })
-                        .then(page => {
-                          page.getDialog().getBody().crop(-50, -50);
-                          page.getDialog().close();
-                          page.getContent().submit();
-                        })
-                        .then(page =>
-                            page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
-                        .then(page =>
-                            page.getContent().getImageHeight().invoke('text').then(CroppedImageHeight => {
-                              expect(CroppedImageHeight).not.equal(OriginalImageHeight);
-                            })
-                        );
-                    });
-                  });
-            });
+            .then(page => page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
+            .then(page => page.getContent().getImageHeight().invoke('text').then(originalImageHeight =>
+                cy.get('@currentPage')
+                  .then(page => page.getContent().openDropDown().openEdit())
+                  .then(page => page.getDialog().getBody().crop(-50, -50))
+                  .then(page => page.getDialog().close())
+                  .then(page => page.getContent().submit())
+                  .then(page => page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
+                  .then(page => page.getContent().getImageHeight().invoke('text').should('not.equal', originalImageHeight))));
       });
 
       it('Rotate image', () => {
         openAssetsPage()
-            .then(page =>
-                page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
-            .then(page => {
-              page.getContent().getFileModifiedDate().invoke('text').then(OriginalModifiedDate => {
+            .then(page => page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
+            .then(page => page.getContent().getFileModifiedDate().invoke('text').then(originalModifiedDate =>
                 cy.get('@currentPage')
-                  .then(page => {
-                    page.getContent().openDropDown().openEdit();
-                  })
-                  .then(page => {
-                    page.getDialog().getBody().rotate();
-                    page.getDialog().close();
-                    page.getContent().submit();
-                  })
-                  .then(page =>
-                      page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
-                  .then(page =>
-                      page.getContent().getFileModifiedDate().invoke('text').then(LastModifiedDate => {
-                        expect(OriginalModifiedDate).not.equal(LastModifiedDate);
-                      })
-                  );
-              });
-            });
+                  .then(page => page.getContent().openDropDown().openEdit())
+                  .then(page => page.getDialog().getBody().rotate())
+                  .then(page => page.getDialog().close())
+                  .then(page => page.getContent().submit())
+                  .then(page => page.getContent().getKebabMenu(testFileInfo.name).openDropdown().openEdit())
+                  .then(page => page.getContent().getFileModifiedDate().invoke('text').should('not.equal', originalModifiedDate))));
       });
+
     });
 
     describe(['ENG-2680'], 'Asset Browsing', () => {
 
       it('Displaying correct item counts when searching with zero results', () => {
         openAssetsPage()
-            .then(page => {
-              page.getContent().openAdvancedFilter();
-              page.getContent().getSearchTextfield().clear();
-              page.getContent().getSearchTextfield().type('z');
-              page.getContent().submitSearch();
-              page.getContent().getFilterResultItemCount().invoke('text').should('be.equal', 'Your search returned 0 results |\n' +
-                  '    Page 0 - 0');
-            });
+            .then(page => page.getContent().openAdvancedFilter())
+            .then(page => page.getContent().getSearchTextfield().then(input => {
+              page.getContent().clear(input);
+              page.getContent().type(input, 'z');
+            }))
+            .then(page => page.getContent().submitSearch())
+            .then(page => page.getContent().getFilterResultItemCount().invoke('text')
+                              .should('be.equal', 'Your search returned 0 results |\n    Page 0 - 0'));
       });
+
     });
 
   });
+
   const testFileInfo = {path: 'upload/image1.JPG', name: 'image1.JPG', type: 'image/jpeg'};
   const testMetadata = {group: 'administrators', categories: [], type: 'image'};
 
-  const openAssetsPage = () => {
-    return cy.get('@currentPage')
-             .then(page => page.getMenu().getContent().open().openAssets());
-  };
+  const openAssetsPage = () => cy.get('@currentPage').then(page => page.getMenu().getContent().open().openAssets());
+
 });
